Instructions
When the page loads in it will prompt you to select a screen, window, or tab. Select what you want to mirror and than it will take you to the content you selected. After that 
go back to the screen with the giant start button and press it. That will create a small video player and you can move it anywhere you want and resize it.